backend/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
fastapi==0.115.5
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
mccabe==0.7.0
PyMySQL==1.1.1
pydantic==2.10.3
pydantic-core==2.27.1
pydantic-settings==2.6.1
pycodestyle==2.12.1
pyflakes==3.2.0
pylint==3.3.1
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.1
watchfiles==1.0.0
websockets==14.1
httpx>=0.24.0
fastapi
uvicorn[standard]
pydantic>=2
pydantic_settings
sqlalchemy
python-dotenv
httpx
numpy
pandas
tqdm
sqlalchemy
alembic
pymysql
cryptography
pydantic-settings
requests
pytest
fastapi
sqlalchemy.orm